requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
binaryornot==0.4.4
black==24.10.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
cloudinary==1.44.1
dj-database-url==3.0.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-modeltranslation==0.19.11
django-unfold==0.34.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
environs==11.2.1
gunicorn==23.0.0
isort==5.13.2
Jinja2==3.1.4
kombu==5.4.2
magic-filter==1.0.12
pillow==11.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tenacity==9.0.0
whitenoise==6.9.0
yarl==1.18.3
# Quyidagilar o'chirildi: click, click-didyoumean, click-pkg, click-plugins, click-repl, colorlog, cookiecutter, environ, frozenlist, g4f, idna, inflection, jsonschema, jsonschema-specifications, jst-aicommit, jst-django, jst-django-core, markdown-it-py, MarkupSafe, marshmallow, mdurl, multidict, mypy-extensions, nest-asyncio, packaging, pathspec, payme-pkg, platformdirs, polib, prompt-toolkit, propcache, pycryptodome, pydantic, pydantic_core, Pygments, PyJWT, pyTelegramBotAPI, questionary, referencing, rich, rpds-py, shellingham, six, text-unidecode, tqdm, typer, types-python-dateutil, typing_extensions, tzdata, uritemplate, urllib3, vine, wcwidth, zipfile36.